update TimeSeriesToTimeHist to TimeSeriesToHist1D
