+++ snewpdag/plugins/ops/TimeSeriesToHist1D.py
@@ -1,0 +1,35 @@
+"""
+TimeSeriesToHist1D - use a TimeSeries to fill a Hist1D.
+
+Arguments:
+  in_field:  input field name
+  out_field:  output field name
+  nbins:  number of bins
+  start:  float to indicate start time
+  stop:  float to indicate stop time
+"""
+import logging
+import numpy as np
+
+from snewpdag.dag import Node
+from snewpdag.values import Hist1D, TimeSeries
+
+class TimeSeriesToHist1D(Node):
+  def __init__(self, in_field, out_field, nbins, start, stop, **kwargs):
+    self.in_field = in_field
+    self.out_field = out_field
+    self.nbins = nbins
+    self.start = start
+    self.stop = stop
+    super().__init__(**kwargs)
+
+  def alert(self, data):
+    if self.in_field in data:
+      ts = data[self.in_field]
+      if isinstance(ts, TimeSeries):
+        th = Hist1D(self.nbins, self.start, self.stop)
+        th.fill(ts.times)
+        data[self.out_field] = th
+        return data
+    return False
+
