+++ snewpdag/plugins/ops/__init__.py
@@ -2,7 +2,7 @@
 from .NewTimeHist import NewTimeHist
 from .NewTimeSeries import NewTimeSeries
 
-from .TimeSeriesToTimeHist import TimeSeriesToTimeHist
+from .TimeSeriesToHist1D import TimeSeriesToHist1D
 
 from .WriteField import WriteField
